requirements.txt
pip
autopep8
Flask
gunicorn
Werkzeug
yfinance
flask_cors
vercel_blob
python-dotenv
fastapi
uvicorn
jinja2
nse
dhanhq
pandas
NorenRestApi
pyotp
azure-storage-blob
azure-identity
azure-ai-documentintelligence
groq